Cleaning up mabs
